+++ lib/screens/tasks/tasks_detail.dart
@@ -4,9 +4,9 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:intl/intl.dart';
 
 import 'package:board/models/task_model.dart';
-import 'package:board/providers/board_route_delegate_state.dart';
-import 'package:board/providers/task_detail_state.dart';
-import 'package:board/providers/tasks_state.dart';
+import 'package:board/providers/board_route_delegate_provider.dart';
+import 'package:board/providers/task_detail_provider.dart';
+import 'package:board/providers/tasks_provider.dart';
 
 class TasksDetailScreen extends HookConsumerWidget {
   TasksDetailScreen({
